Prompt updates
- prompt, added a space before the coloring, plus I added the color blue
if I want to use it in the future.
- .bashrc, I have a short explanation in the file but here is some more.
I tried using .inputrc to set vim keys for the terminal and also have
the modes show up. From my conclution to why it didn't work I added bind
commands together with set to achieve my goal of the whole thing.
I believe that it's because I add my custom prompt after the fact that
.inputrc is read which means I'm overwriting the addition.