Install script update
- Added a part to the colors section in the install script.
Previously we have had the dircolors part only, printing to the
.dircolors file. Now we use sed to change the colors of directories to
yellow instead of the traditional blue.

The reasoning is that I'm using the gruvbox colorscheme for basically
everything and the file links and directories looks so alike that I need
to separate it in some way and changing directory colors to yellow feels
like a perfect match.